Integer dt point

+++ app/src/main/java/com/worldventures/dreamtrips/modules/profile/view/cell/ProfileCell.java
@@ -319,7 +319,7 @@ public class ProfileCell extends AbstractCell<User> implements Expandable {
     }
 
     private void setDreamTripPoints(String count) {
-        dtPoints.setText(Html.fromHtml(context.getString(R.string.profile_dt_points, count)));
+        dtPoints.setText(Html.fromHtml(context.getString(R.string.profile_dt_points, Float.valueOf(count).intValue())));
     }
 
     private void setIsFriend(boolean isFriend) {
